bug fix for returning result early

+++ soar/Core/SoarKernel/src/extension/epmem_manager.h
@@ -45,7 +45,7 @@ private:
 	int msgCount;
     int id;
 	
-	int totalEpCnt;
+	double totalEpCnt;
     int windowSize;
     int currentSize;
 	bool sendEpNextTime;
